backend/requirements.txt
flask
python-nmap
netifaces